AP_GPS: move blended-GPS functions into AP_GPS_Blended
collects all of these together in preparation for making a backend
